requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.6
python-dotenv==1.0.1
google-generativeai==0.8.0
openai==1.50.0
chromadb==0.5.0
numpy<2
PyMuPDF==1.24.0
apscheduler==3.10.4
pydantic==2.9.0
